// src/render/core/types.ts
export interface RenderStats {
  fps: number;
  frameTime: number;
  drawCalls: number;
  instancedDrawCalls: number;
  computePasses: number;
  triangles: number;
  vertices: number;
  bufferMemory: number;
  textureMemory: number;
  gpuTime: number;
  shadowMapTime: number;
  postProcessingTime: number;
  pipelineChanges: number;
  textureBindings: number;
  bufferUpdates: number;
}